arch: common: semihost: add semihosting operations
Add an API that utilizes the ARM semihosting mechanism to interact with
the host system when a device is being emulated or run under a debugger.

RISCV is implemented in terms of the ARM implementation, and therefore
the ARM definitions cross enough architectures to be defined 'common'.

Functionality is exposed as a separate API instead of syscall
implementations (`_lseek`, `_open`, etc) due to various quirks with
the ARM mechanisms that means function arguments are not standard.

For more information see:

impl

https://developer.arm.com/documentation/dui0471/m/what-is-semihosting-
Signed-off-by: Jordan Yates <jordan.yates@data61.csiro.au>
GitOrigin-RevId: d8f186aa4a18ab2230a9733c3a133fe1433aa87d
Change-Id: I378025028ff29eee2e2d6af553ee9f1906eb4dac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3602409
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>