net: if: Connect interface with offloaded socket implementation
Instead of keeping a boolean informing whether a network interface is
offloaded at socket layer or not, keep a pointer to a function which
allows to create an offloaded socket. Native interfaces keep this as
NULL, while for offloaded interfaces it allows to connect an offloaded
socket implementation with an interface.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>
GitOrigin-RevId: fa8ba73833587ebb0133c70ddeb2992a85939128
Change-Id: Ib387d439af94e6f7dbf342edbab721f3cba1a31e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3596505
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>