boards: arm: stm32: set LEDs period to 20ms
Most LEDs had 0 or 4 nanoseconds set as a period, a value that doesn't
make sense for a PWM signal driving an LED. A period of 20ms (50Hz) is a
frequently used value as it is above the flicker fusion threshold. All
STM32 based boards have been updated.

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>
GitOrigin-RevId: 5a0b12c2303537507db1ae4700f47eb79f49014f
Change-Id: Icebae41f596ffb109849ae9e641e8afee3f52129
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3575817
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>