ztest: if runid is set, propagate it
define test run id and make it available to ztest. This will be used to
verify we are evaluating the same test we have just built and flashed
and that we are not looking at some old output.

Existing code in ztest will use this, and twister will define the run
id.

Signed-off-by: Anas Nashif <anas.nashif@intel.com>
GitOrigin-RevId: 1537ee7a0bf44979d4ed1c27d0d8b961cf4d3f80
Change-Id: Ia34e0ea99cfbfe2be5f0a8b545d9a5a29350f530
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3580596
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>