twister: generate and verify test run id
Set run id (md5 of instance and a random number) for each test and match
what was set during build with the output on the screen. This will
verify that we are evaluating the same test we have uploaded and not
some previous output on the screen.

Signed-off-by: Anas Nashif <anas.nashif@intel.com>
GitOrigin-RevId: 540327ee4fe5730fecef500db6afde8833e8efb7
Change-Id: I23796b72c9fcb7c028020aa40d90c2b681c6e667
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3585382
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>