scripts: logging/dictionary/sys-t: log module info in collateral
This updates the MIPI Sys-T collateral generation to include log
modules in the XML file. This allows the parser to show where
each log message is coming from.

Note that the addition of Name propety in sys:Client is to clarify
that this collateral is from Zephyr.

Also note he addition of Mask property in syst:Guid is for parser
to match this collateral to any incoming messages, as the module
IDs are encoded in the set bits in the GUID. Without the mask,
the parser will discard any messages with non-zero module IDs.
This is based on the observation of the reference parser,
systprint, where pseudo GUIDs are generated from the module/unit
IDs (origin unit) in the incoming messages for matching with
collaterals.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>
GitOrigin-RevId: 2e2c8cc931d2db9d02e2226438577a85b8f75ead
Change-Id: I802f34b4881f7786dff6e0b518cd817cb2622e07
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3637683
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>