x86: qemu: add a newline after "Booting from ROM.."
Under QEMU and SeaBIOS, everything gets to be printed
immediately after "Booting from ROM.." as there is no newline.
This prevents parsing QEMU console output for the very first
line where it needs to match from the beginning of the line.
So add a dummy newline here so the next output is at
the beginning of a line.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>
GitOrigin-RevId: 7a431dca95eba22c966eb30cf7798aacd50c500b
Change-Id: I5ebcd3c00af20b1a2e64f42053bde800594b7311
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3580463
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>