drivers: wifi: esp_at: handle external reset sources
Add an option that signifies that the ESP modem may be reset at the same
time as the SoC by an external source. When this is the case, we first
wait for an unsolicited "ready" message from the modem, before
attempting to reset the device. This prevents two initialisation
sequences attempting to run at the same time.

We still want to wait for the complete initialisation sequence to
complete before returning in this case.

(cherry picked from commit 21ed808ba1d1f16ac43e617261ce278a04226f12)

Original-Signed-off-by: Jordan Yates <jordan.yates@data61.csiro.au>
GitOrigin-RevId: 21ed808ba1d1f16ac43e617261ce278a04226f12
Change-Id: I3e4b6559957823573c5662047394ac1dcc54c512
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4668929
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>