kernel: mutex: remove unnecessary schedule locking
Removes an unnecessary schedule lock/unlock pair from k_mutex_unlock().

mutex owner) will be able to modify the mutex object AND that a
recursive unlock ought never trigger any reschedule (as it does not
touch the pend queue), then performing a schedule lock is not needed
prior to testing for a recursive unlock.

Furthermore, even if it is not a recursive unlock, then a schedule lock
is superfluous as the existing spinlock provides sufficient protection.

Rationale: Given that only the current thread (which would also be the
Signed-off-by: Peter Mitsis <peter.mitsis@intel.com>
GitOrigin-RevId: e9b288b7137e5fb6ded18334aaa5a5f9c23ebb17
Change-Id: I3a3e61dda9c86ab8884169bb1422d378e73d602c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3599718
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>