modules: canopennode: can_get_max_filters() is optional
Support for the can_get_max_filters() API function is optional and may
not be supported by all drivers.

The check for an adequate amount of filters is just there in order to
fail early. If the driver does not support can_get_max_filters(),
reporting of a failure will be delayed until a filter addition fails.

Signed-off-by: Henrik Brix Andersen <hebad@vestas.com>
GitOrigin-RevId: 38ee9535473a0bc16bbe2b69a7d63c814e67e15a
Change-Id: I6388e98b284077a25b0c43588e0de27c038df316
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3593662
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>