doc: Rearrange content of section "Labeling issues and pull requests"
Mostly due to limited support for tables in rst format, information on
labels used for Zephyr issues and pull requests in Github is hard to
navigate and is visually not pleasing. Rearange the content of the
section so that bullets are mainly utilized rather than tables. Format
label names as 'emphasis', rather that 'monospace' (which should be
used primarily for code snippets, and Github labels are not code).
Also, group the information according to applicability (taking into
account if a label is applicable to:

  * issues only;
  * pull requests only;
  * both issues and pull requests.

Label names and decriptions are otherwise left 'as is' in this commit.

(cherry picked from commit 3f1e43b9c77fd91dbc409716861d70dc00e04714)

Original-Signed-off-by: Aleksandar Markovic <aleksandar.markovic.sa@gmail.com>
GitOrigin-RevId: 3f1e43b9c77fd91dbc409716861d70dc00e04714
Change-Id: I0e9ea19a131f347a6248e140c0cc46253c710b5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3734927
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>