kernel: update k_thread_state_str() API
When threads are in more than one state at a time, k_thread_state_str()
returns a string that lists each of its states delimited by a '+'.
This in turn necessitates a change to the API that includes both a
pointer to the buffer to use for the string and the size of the buffer.

Signed-off-by: Peter Mitsis <peter.mitsis@intel.com>
GitOrigin-RevId: a30cf39975b049d4630b86c986e2974d67da1d1b
Change-Id: I53d148ab6bd81885c056e4cdad05cd4c1975d937
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3599716
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>